Add new github action to test khiops installation
The test-khiops-install checks if the output of Khiops in the console is as expected:
- no mpi or java errors
- the number of process launched is KHIOPS_PROC_NUMBER

The proc number is set to 4 while testing khiops installation. The default value is the
physical proc number, then 2 proc on the github runner. With only 2 procs, khiops is
runing in serial without call to mpiexec.